+ui true commit im dumb

+++ backend/src/main/java/com/example/backend/dto/MlReturnDto.java
@@ -7,8 +7,9 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
+
 public class MlReturnDto {
     int id;
-    double similarity;
+    float similarity;
 
 }
